Characters are not game components moved them
